Reorganise sidebar folder structure
