+++ src/components/Viewer.vue
@@ -84,7 +84,7 @@
 
     </div>
 
-    <sidebar
+    <task-sidebar
       v-if="currentTask"
       :task="currentTask"
       :showNote="showNote"
@@ -108,7 +108,7 @@
         @inputfocus="onTranscribeInputFocus"
         @inputblur="onTranscribeInputBlur">
       </transcribe-sidebar-item>
-    </sidebar>
+    </task-sidebar>
 
     <selector
       v-if="selectorEnabled"
@@ -140,10 +140,10 @@ import BrowseModal from '@/components/modals/Browse'
 import ToolbarControls from '@/components/controls/Toolbar'
 import PanControls from '@/components/controls/Pan'
 import ZoomControls from '@/components/controls/Zoom'
-import Sidebar from '@/components/sidebar/Sidebar'
-import NavigationSidebar from '@/components/sidebar/Navigation'
-import SelectSidebarItem from '@/components/sidebar/items/Select'
-import TranscribeSidebarItem from '@/components/sidebar/items/Transcribe'
+import TaskSidebar from '@/components/sidebars/Task'
+import NavigationSidebar from '@/components/sidebars/Navigation'
+import SelectSidebarItem from '@/components/sidebars/items/Select'
+import TranscribeSidebarItem from '@/components/sidebars/items/Transcribe'
 import Selector from '@/components/Selector'
 import Task from '@/model/Task'
 import Annotator from '@/model/Annotator'
@@ -274,7 +274,7 @@ export default {
     ToolbarControls,
     PanControls,
     ZoomControls,
-    Sidebar,
+    TaskSidebar,
     NavigationSidebar,
     Selector,
     SelectSidebarItem,
